Blue line indicator under hymn lyrics bar
