+++ Hymns/Common/TabView/TabView.swift
@@ -10,18 +10,20 @@ public struct IndicatorTabView<TabType: TabItem>: View {
     @Binding var currentTab: TabType
     let tabItems: [TabType]
 
+
     public var body: some View {
         return
             VStack(alignment: .center) {
-                Rectangle()
-                    .foregroundColor(.white)
-                    .overlay(currentTab.content)
 
                 Rectangle()
                     .shadow(radius: 0, y: -0.2)
                     .frame(height: 50)
                     .foregroundColor(.white)
                     .overlay(TabBar(currentTab: $currentTab, tabItems: tabItems))
+                
+                Rectangle()
+                    .foregroundColor(.white)
+                    .overlay(currentTab.content)
         }
     }
 }
@@ -53,3 +55,4 @@ struct IndicatorTabView_Previews: PreviewProvider {
         }
     }
 }
+
